Recomentations: fixed asynchronous error
Removed use of addMega and mountLoad.
Manually wrapped recommendations component with a loading message until the data arrives.
Implemented onServer method so ss-resolve will wait for the data.
Used the callback edition of setState per react docs (when relying on previous state).
Moved binding to constructor.
Async error was: the timer was cleared instantly because the data hadn't returned yet.